Split project.dart into CMake and Xcode projects
