+++ packages/flutter_tools/lib/src/linux/application_package.dart
@@ -6,8 +6,8 @@ import '../application_package.dart';
 import '../base/file_system.dart';
 import '../build_info.dart';
 import '../cmake.dart';
+import '../cmake_project.dart';
 import '../globals_null_migrated.dart' as globals;
-import '../project.dart';
 
 abstract class LinuxApp extends ApplicationPackage {
   LinuxApp({required String projectBundleId}) : super(id: projectBundleId);
