+++ packages/flutter_tools/lib/src/linux/build_linux.dart
@@ -12,11 +12,11 @@ import '../base/utils.dart';
 import '../build_info.dart';
 import '../cache.dart';
 import '../cmake.dart';
+import '../cmake_project.dart';
 import '../convert.dart';
 import '../flutter_plugins.dart';
 import '../globals_null_migrated.dart' as globals;
 import '../migrations/cmake_custom_command_migration.dart';
-import '../project.dart';
 
 // Matches the following error and warning patterns:
 // - <file path>:<line>:<column>: (fatal) error: <error...>
