+++ packages/flutter_tools/lib/src/windows/application_package.dart
@@ -11,8 +11,8 @@ import '../base/file_system.dart';
 import '../base/utils.dart';
 import '../build_info.dart';
 import '../cmake.dart';
+import '../cmake_project.dart';
 import '../globals_null_migrated.dart' as globals;
-import '../project.dart';
 
 abstract class WindowsApp extends ApplicationPackage {
   WindowsApp({@required String projectBundleId}) : super(id: projectBundleId);
