+++ packages/flutter_tools/lib/src/windows/build_windows.dart
@@ -14,11 +14,11 @@ import '../base/utils.dart';
 import '../build_info.dart';
 import '../cache.dart';
 import '../cmake.dart';
+import '../cmake_project.dart';
 import '../convert.dart';
 import '../flutter_plugins.dart';
 import '../globals_null_migrated.dart' as globals;
 import '../migrations/cmake_custom_command_migration.dart';
-import '../project.dart';
 import 'install_manifest.dart';
 import 'visual_studio.dart';
